+++ Runtime/Models/AgreementModels.cs
@@ -1,4 +1,4 @@
-// Copyright (c) 2020 - 2023 AccelByte Inc. All Rights Reserved.
+// Copyright (c) 2020 - 2024 AccelByte Inc. All Rights Reserved.
 // This is licensed software from AccelByte Inc, for limitations
 // and restrictions contact your company contract manager.
 using System.Runtime.Serialization;
@@ -100,6 +100,7 @@ namespace AccelByte.Models
     public class AcceptAgreementResponse
     {
         [DataMember] public bool proceed;
+        [DataMember(Name = "comply")] public bool Comply;
     }
 
     [DataContract, Preserve]